♻️ E5E8EF blueprints: decompose validation helpers
