+++ .agentplane/tasks/202605290203-E5E8EF/pr/github-body.md
@@ -1,0 +1,33 @@
+Task: `202605290203-E5E8EF`
+Title: Blueprint validate decomposition
+Canonical task record: `.agentplane/tasks/202605290203-E5E8EF/README.md`
+
+## Summary
+
+Blueprint validate decomposition
+
+Decompose packages/agentplane/src/blueprints/validate.ts by extracting shared graph/evidence validation helpers while preserving blueprint and plan validation behavior and reducing runtime hotspot warnings.
+
+## Scope
+
+- In scope: Decompose packages/agentplane/src/blueprints/validate.ts by extracting shared graph/evidence validation helpers while preserving blueprint and plan validation behavior and reducing runtime hotspot warnings.
+- Out of scope: unrelated refactors not required for "Blueprint validate decomposition".
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T02:04:06.763Z
+- Branch: task/202605290203-E5E8EF/blueprint-validate-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
