+++ .agentplane/tasks/202605290203-E5E8EF/pr/github-title.txt
@@ -1,0 +1,1 @@
+🚧 E5E8EF task: Blueprint validate decomposition [202605290203-E5E8EF]
