+++ .agentplane/tasks/202605290203-E5E8EF/pr/review.md
@@ -1,0 +1,36 @@
+# PR Review
+
+Created: 2026-05-29T02:04:06.763Z
+
+## Task
+
+- Task: `202605290203-E5E8EF`
+- Title: Blueprint validate decomposition
+- Status: DOING
+- Branch: `task/202605290203-E5E8EF/blueprint-validate-decomposition`
+- Canonical task record: `.agentplane/tasks/202605290203-E5E8EF/README.md`
+
+## Verification
+
+- State: pending
+- Note: Not recorded yet.
+- Canonical workflow state lives in the task README.
+
+## Handoff Notes
+
+- No handoff notes recorded yet. Use `agentplane pr note ...` to append one.
+
+<!-- BEGIN AUTO SUMMARY -->
+<details>
+<summary>Raw evidence</summary>
+
+- Updated: 2026-05-29T02:04:06.763Z
+- Branch: task/202605290203-E5E8EF/blueprint-validate-decomposition
+- Head: computed live by `agentplane pr check` / `agentplane integrate`
+
+```text
+No changes detected.
+```
+
+</details>
+<!-- END AUTO SUMMARY -->
